pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nyancobot"
version = "0.2.0"
description = "Secure, production-ready AI agent framework forked from nanobot with enhanced security, multi-agent collaboration, and automation features"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["ai", "agent", "llm", "automation", "security", "multi-agent", "playwright", "mcp"]
authors = [
    {name = "nyancobot contributors"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "playwright>=1.40.0",
    "litellm>=1.30.0",
    "fastmcp>=0.1.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
]
slack = ["slack-bolt>=1.18"]
discord = ["discord.py>=2.3"]
line = ["line-bot-sdk>=3.0"]
whatsapp = ["httpx>=0.27"]
webhook = ["fastapi>=0.110", "uvicorn>=0.29"]
all-channels = ["slack-bolt>=1.18", "discord.py>=2.3", "line-bot-sdk>=3.0", "httpx>=0.27", "fastapi>=0.110", "uvicorn>=0.29"]

[project.urls]
Homepage = "https://github.com/asiokun/nyancobot"
Repository = "https://github.com/asiokun/nyancobot"
Issues = "https://github.com/asiokun/nyancobot/issues"
Documentation = "https://github.com/asiokun/nyancobot#readme"

[tool.setuptools]
packages = ["nyancobot", "nyancobot.channels", "nyancobot.agent", "nyancobot.agent.tools", "nyancobot.agent.tools.browser_stealth", "nyancobot.providers"]

[tool.setuptools.package-data]
nyancobot = ["py.typed"]
"nyancobot.agent.tools.browser_stealth" = ["stealth.js"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
